microservices/collector/requirements.txt
fastapi
fastapi-pagination[sqlalchemy]
uvicorn
trendspy
sqlalchemy
apscheduler
python-dotenv
PyYAML
pandas
fastapi-events
pydantic-settings 
httpx
Tenacity
python-consul